Completed day 09 in C++ as well.
Required changing IntCode to run on long instead of int, which
required changes to earlier C++ solutions as well.

+++ cpp/d02.cc
@@ -3,7 +3,7 @@
 #include "ReadInputs.hh"
 #include "IntCode.hh"
 
-int apply_func(int a, int b, std::vector<int> memory) {
+int apply_func(int a, int b, std::vector<long> memory) {
   memory[1] = a;
   memory[2] = b;
