+++ cpp/include/ReadInputs.hh
@@ -7,6 +7,6 @@
 std::string read_file(std::string filename);
 std::vector<std::string> read_lines(std::string filename);
 
-std::vector<int> read_integers(std::string filename);
+std::vector<long> read_integers(std::string filename);
 
 #endif /* _READINPUTS_H_ */
